some fix for dto transazione

+++ src/main/java/aldovalzani/capstone_be/dto/TransazioneDTO.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull;
 public record TransazioneDTO(
         @NotNull long idWallet,
         @NotNull long idCrypto,
-        @NotNull long quantita,
-        @NotNull long prezzo,
+        @NotNull double quantita,
+        @NotNull double prezzo,
         @NotBlank String tipoTransazione
 ) {}
